requirements.txt
beautifulsoup4 >= 4.5.1, < 5.0
click >= 6.6, < 7.0
pyyaml >= 3.11, < 4.0
requests >= 2.11.1, < 3.0
requests-kerberos >= 0.10.0, < 1.0
atomicwrites >= 1.1.5, < 2.0
